+++ ios/Classes/Banner/BannerAdController.swift
@@ -4,11 +4,11 @@ import GoogleMobileAds
 class BannerAdController: NSObject {
 
     var bannerView: GADBannerView!
-
-//    var loadRequested: ((MethodChannel.Result) -> Unit)? = null
-
+    var loadRequested: (() -> Void)? = nil
     let id: String
     let channel: FlutterMethodChannel
+    var result : FlutterResult?=nil
+
 
     init(id: String, channel: FlutterMethodChannel) {
         self.id = id
@@ -19,11 +19,13 @@ class BannerAdController: NSObject {
     }
 
     private func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
+        self.result=result
         _ = call.arguments as? [String: Any]
 
         switch call.method {
             case "loadAd":
                 channel.invokeMethod("loading", arguments: nil)
+                if(loadRequested != nil){loadRequested!() }
             default:
                 result(FlutterMethodNotImplemented)
         }