+++ ios/Classes/Extensions/Int+dp.swift
@@ -4,5 +4,11 @@ extension Int {
         let density = UIScreen.main.scale
         return Int(CGFloat(self) * density)
     }
+    
+    func dpReverse() -> Int {
+        if (self == -1 || self == -2) {return self}
+        let density = UIScreen.main.scale
+        return Int(CGFloat(self) / density)
+    }
 }