+++ ios/Classes/Rewarded/RewardedAdController.swift
@@ -27,10 +27,17 @@ class RewardedAdController: NSObject,GADFullScreenContentDelegate {
         case "loadAd":
             channel.invokeMethod("loading", arguments: nil)
             let unitId: String = params?["unitId"] as! String
-            GADRewardedAd.load(withAdUnitID: unitId, request: GADRequest()) { (ad : GADRewardedAd?, error:Error?) in
+            let request = GADRequest()
+            if #available(iOS 13.0, *) {
+                request.scene = UIApplication.shared.keyWindow?.windowScene
+            }
+            GADRewardedAd.load(withAdUnitID: unitId, request: request) { (ad : GADRewardedAd?, error:Error?) in
                 if error != nil {
                     self.rewardedAd = nil
-                    self.channel.invokeMethod("onAdFailedToLoad", arguments: error.debugDescription)
+                    self.channel.invokeMethod("onAdFailedToLoad", arguments: [
+                        "errorCode": (error! as NSError).code,
+                        "message": (error! as NSError).localizedDescription
+                    ])
                     result(false)
                 }
                 else{