+++ lib/src/app_open/app_open.dart
@@ -141,7 +141,7 @@ class AppOpenAd extends LoadShowAd<FullScreenAdEvent> {
     /// Force to load an ad even if another is already avaiable
     bool force = false,
 
-    /// The timeout of this ad. If null, defaults to `Duration(seconds: 30)`
+    /// The timeout of this ad. If null, defaults to 1 minute
     Duration? timeout,
   }) async {
     ensureAdNotDisposed();
@@ -180,7 +180,7 @@ class AppOpenAd extends LoadShowAd<FullScreenAdEvent> {
     }).timeout(
       timeout ?? this.loadTimeout,
       onTimeout: () {
-        if (!onEventController.isClosed)
+        if (!onEventController.isClosed && !isLoaded)
           onEventController.add({
             FullScreenAdEvent.loadFailed: AdError.timeoutError,
           });