+++ lib/src/banner/controller.dart
@@ -227,7 +227,7 @@ class BannerAdController extends LoadShowAd<BannerAdEvent>
     /// Force to load an ad even if another is already avaiable
     bool force = false,
 
-    /// The timeout of this ad. If null, defaults to 30 seconds
+    /// The timeout of this ad. If null, defaults to 1 minute
     Duration? timeout,
   }) async {
     ensureAdNotDisposed();
@@ -236,7 +236,7 @@ class BannerAdController extends LoadShowAd<BannerAdEvent>
     isLoaded = (await channel.invokeMethod<bool>('loadAd').timeout(
       timeout ?? this.loadTimeout,
       onTimeout: () {
-        if (!onEventController.isClosed)
+        if (!onEventController.isClosed && !isLoaded)
           onEventController.add({
             BannerAdEvent.loadFailed: AdError.timeoutError,
           });