+++ lib/src/native/controller/controller.dart
@@ -334,7 +334,7 @@ class NativeAdController extends LoadShowAd<NativeAdEvent>
     /// Force to load an ad even if another is already avaiable
     bool force = false,
 
-    /// The timeout of this ad. If null, defaults to `Duration(seconds: 30)`
+    /// The timeout of this ad. If null, defaults to 1 minute
     Duration? timeout,
   }) async {
     ensureAdNotDisposed();
@@ -347,7 +347,7 @@ class NativeAdController extends LoadShowAd<NativeAdEvent>
     }).timeout(
       timeout ?? this.loadTimeout,
       onTimeout: () {
-        if (!onEventController.isClosed)
+        if (!onEventController.isClosed && !isLoaded)
           onEventController.add({
             NativeAdEvent.loadFailed: AdError.timeoutError,
           });