+++ lib/src/native/native_ad_widget.dart
@@ -132,7 +132,7 @@ class NativeAd extends StatefulWidget {
 
   /// The ad will stop loading after a specified time.
   ///
-  /// If `null`, defaults to `Duration(seconds: 30)`
+  /// If `null`, defaults to 1 minute
   final Duration? loadTimeout;
 
   /// Use hybrid composition in this ad. This has effect only on Android