+++ lib/src/native/utils.dart
@@ -75,7 +75,8 @@ class AdBorderRadius {
 
   /// Creates a vertically symmetric border radius where the top
   /// and bottom sides of the rectangle have the same value.
-  static AdBorderRadius vertical({double? top, double? bottom}) => AdBorderRadius(
+  static AdBorderRadius vertical({double? top, double? bottom}) =>
+      AdBorderRadius(
         topLeft: top,
         topRight: top,
         bottomLeft: bottom,